Changed to float to allow decimals.

+++ main.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 int main()
 {
-    int x;
+    float x;
 
     cout << "Enter a value: ";
     cin >> x;
@@ -16,5 +16,5 @@ int main()
     return 0;
 }
 
-//Limit is 2147483647
-//Limit for unsigned 4294967295
+//In  3.14159, 2.9e38,   2.9e39,       5.1e-38,  5.1e-39,  0.0000000005
+//Out 3.14159, 2.9e+038, 3.40282e+038, 5.1e-038. 5.1e-039, 5e-010
